refactor(server): rework test hooks
drop db migration before and after hooks, add helpers and root hooks

fix #531
