+++ packages/server/modules/core/tests/commits.spec.js
@@ -1,29 +1,21 @@
 /* istanbul ignore file */
-const chai = require( 'chai' )
-const chaiHttp = require( 'chai-http' )
-const assert = require( 'assert' )
+const expect = require( 'chai' ).expect
 
 const appRoot = require( 'app-root-path' )
-const { init } = require( `${appRoot}/app` )
-const knex = require( `${appRoot}/db/knex` )
-
-const expect = chai.expect
-chai.use( chaiHttp )
+const { beforeEachContext } = require( `${appRoot}/test/hooks` )
 
 
 const { createUser } = require( '../services/users' )
-const { createStream, getStream, updateStream, deleteStream } = require( '../services/streams' )
-const { createObject, createObjects } = require( '../services/objects' )
+const { createStream  } = require( '../services/streams' )
+const { createObject } = require( '../services/objects' )
 const { createBranch } = require( '../services/branches' )
 
 const {
   createCommitByBranchName,
-  createCommitByBranchId,
   updateCommit,
   getCommitById,
   deleteCommit,
   getCommitsTotalCountByBranchName,
-  getCommitsByBranchId,
   getCommitsByBranchName,
   getCommitsByStreamId,
   getCommitsTotalCountByStreamId,
@@ -32,7 +24,6 @@ const {
 } = require( '../services/commits' )
 
 describe( 'Commits @core-commits', ( ) => {
-
   let user = {
     name: 'Dimitrie Stefanescu',
     email: 'didimitrie4342@gmail.com',
@@ -60,10 +51,7 @@ describe( 'Commits @core-commits', ( ) => {
   }
 
   before( async ( ) => {
-    await knex.migrate.rollback( )
-    await knex.migrate.latest( )
-
-    await init()
+    await beforeEachContext( )
 
     user.id = await createUser( user )
     stream.id = await createStream( { ...stream, ownerId: user.id } )
@@ -71,10 +59,6 @@ describe( 'Commits @core-commits', ( ) => {
     testObject.id = await createObject( stream.id, testObject )
     testObject2.id = await createObject( stream.id, testObject2 )
     testObject3.id = await createObject( stream.id, testObject3 )
-  } )
-
-  after( async ( ) => {
-    await knex.migrate.rollback( )
   } )
 
   let commitId1, commitId2, commitId3
@@ -103,7 +87,6 @@ describe( 'Commits @core-commits', ( ) => {
 
     let res = await deleteCommit( { id: tempCommit } )
     expect( res ).to.equal( 1 )
-
   } )
 
   it( 'Should get a commit by id', async ( ) => {
