+++ packages/server/modules/emails/tests/verifications.spec.js
@@ -1,31 +1,21 @@
 /* istanbul ignore file */
-const chai = require( 'chai' )
-const chaiHttp = require( 'chai-http' )
+const expect = require( 'chai' ).expect
 
 const appRoot = require( 'app-root-path' )
-const { init, startHttp } = require( `${appRoot}/app` )
+const { beforeEachContext } = require( `${appRoot}/test/hooks` )
 const request = require( 'supertest' )
-
-const expect = chai.expect
-chai.use( chaiHttp )
 
 const knex = require( `${appRoot}/db/knex` )
 
 const { createUser } = require( `${appRoot}/modules/core/services/users` )
-const {
-  createPersonalAccessToken,
-} = require( `${appRoot}/modules/core/services/tokens` )
+const { createPersonalAccessToken } = require( `${appRoot}/modules/core/services/tokens` )
 
-const {
-  sendEmailVerification,
-} = require( `${appRoot}/modules/emails/services/verification` )
+const { sendEmailVerification } = require( `${appRoot}/modules/emails/services/verification` )
 
 const Verifications = () => knex( 'email_verifications' )
 
 describe( 'Email verifications @emails', () => {
-  let expressApp
-
-
+  let app
 
   const userA = {
     'name': 'd1',
@@ -39,20 +29,12 @@ describe( 'Email verifications @emails', () => {
   }
 
   before( async () => {
-    await knex.migrate.rollback()
-    await knex.migrate.latest()
-
-    let { app } = await init()
-    expressApp = app
+    ( { app } = await beforeEachContext( ) )
 
     userA.id = await createUser( userA )
     userA.token = `Bearer ${( await createPersonalAccessToken( userA.id, 'test token user A', [ 'server:setup', 'streams:read', 'streams:write', 'users:read', 'users:email', 'tokens:write', 'tokens:read', 'profile:read', 'profile:email' ] ) )}`
     userB.id = await createUser( userB )
     userB.token = `Bearer ${( await createPersonalAccessToken( userB.id, 'test token user B', [ 'streams:read', 'streams:write', 'users:read', 'users:email', 'tokens:write', 'tokens:read', 'profile:read', 'profile:email' ] ) )}`
-  } )
-
-  after( async () => {
-    await knex.migrate.rollback()
   } )
 
   describe( 'Create email verification', () => {
@@ -70,7 +52,7 @@ describe( 'Email verifications @emails', () => {
 
       await Verifications( ).where( { id: ver.id } ).del()
 
-      await request( expressApp )
+      await request( app )
         .post( '/auth/emailverification/request' )
         .send( { email: userA.email } )
         .set( 'Authorization', userA.token )
@@ -78,19 +60,19 @@ describe( 'Email verifications @emails', () => {
     } )
     it( 'Should fail to send verification unauthenticated / not the same user',
       async () => {
-        await request( expressApp )
+        await request( app )
           .post( '/auth/emailverification/request' )
           .send( { email: userA.email } )
           .expect( 403 )
 
-        await request( expressApp )
+        await request( app )
           .post( '/auth/emailverification/request' )
           .send( { email: userA.email } )
           .set( 'Authorization', userB.token )
           .expect( 403 )
       } )
     it( 'Should not create a new verification while the previous is valid', async () => {
-      await request( expressApp )
+      await request( app )
         .post( '/auth/emailverification/request' )
         .send( { email: userA.email } )
         .set( 'Authorization', userA.token )
@@ -100,7 +82,7 @@ describe( 'Email verifications @emails', () => {
       await Verifications().where( { email: userA.email } )
         .update( { createdAt: new Date( 0 ) } )
 
-      await request( expressApp )
+      await request( app )
         .post( '/auth/emailverification/request' )
         .send( { email: userA.email } )
         .set( 'Authorization', userA.token )
@@ -109,16 +91,16 @@ describe( 'Email verifications @emails', () => {
   } )
   describe( 'Use email verification', () => {
     it( 'Should not verify without a token', async () => {
-      await request( expressApp )
+      await request( app )
         .get( '/auth/verifyemail' )
         .expect( 400 )
       
-      await request( expressApp )
+      await request( app )
         .get( '/auth/verifyemail?t=' )
         .expect( 400 )
     } )
     it( 'Should not verify with an invalid token', async () => {
-      await request( expressApp )
+      await request( app )
         .get( '/auth/verifyemail?t=bogus' )
         .expect( 404 )
     } )
@@ -132,7 +114,7 @@ describe( 'Email verifications @emails', () => {
       }
       await Verifications().insert( expiredVerification )
 
-      await request( expressApp )
+      await request( app )
         .get( `/auth/verifyemail?t=${expiredVerification.id}` )
         .expect( 400 )
 
@@ -143,7 +125,7 @@ describe( 'Email verifications @emails', () => {
       let verifications = await Verifications().where( { email: userA.email } )
       expect( verifications ).to.have.lengthOf( 2 )
 
-      await request( expressApp )
+      await request( app )
         .get( `/auth/verifyemail?t=${verifications[1].id}` )
         .expect( 302 )
