+++ packages/server/modules/serverinvites/tests/serverInvites.spec.js
@@ -1,26 +1,18 @@
 /* istanbul ignore file */
-const crs = require( 'crypto-random-string' )
-const chai = require( 'chai' )
-const request = require( 'supertest' )
+const expect = require( 'chai' ).expect
 const assert = require( 'assert' )
 const appRoot = require( 'app-root-path' )
-const { async } = require( 'crypto-random-string' )
-const { init, startHttp } = require( `${appRoot}/app` )
 
-const expect = chai.expect
 
-const knex = require( `${appRoot}/db/knex` )
 const { createUser } = require( `${appRoot}/modules/core/services/users` )
 
 const { createAndSendInvite, getInviteById, getInviteByEmail, validateInvite, useInvite, sanitizeMessage } = require( `${appRoot}/modules/serverinvites/services` )
-const { createStream, getStream, getStreamUsers, getUserStreams } = require( `${appRoot}/modules/core/services/streams` )
+const { createStream, getUserStreams } = require( `${appRoot}/modules/core/services/streams` )
 const { createPersonalAccessToken } = require( `${appRoot}/modules/core/services/tokens` )
 
-let serverAddress
+const { beforeEachContext, initializeTestServer } = require( `${appRoot}/test/hooks` )
 
 describe( 'Server Invites @server-invites', ( ) => {
-  let myApp
-
   describe( 'Services @server-invites-services', () => {
     let actor = {
       name: 'Dimitrie Stefanescu',
@@ -29,16 +21,8 @@ describe( 'Server Invites @server-invites', ( ) => {
     }
 
     before( async() => {
-      await knex.migrate.rollback( )
-      await knex.migrate.latest( )
-
-      let { app } = await init()
-      myApp = app
+      await beforeEachContext( )
       actor.id = await createUser( actor )
-    } )
-
-    after( async() => {
-      await knex.migrate.rollback( )
     } )
 
     it( 'should create an invite', async() => {
@@ -179,36 +163,26 @@ describe( 'Server Invites @server-invites', ( ) => {
 
   // TODO: reinstate these tests; not sure why they pass locally and fail on CI
   describe( 'API @server-invites-api', () => {
+    let server, sendRequest
     let actor = {
       name: 'Dimitrie Stefanescu',
       email: 'didimitrie-10000@gmail.com',
       password: 'wtfwtfwtf'
     }
 
-    let testServer, testToken
+    let testToken
 
     before( async() => {
-      await knex.migrate.rollback( )
-      await knex.migrate.latest( )
-
-      let { app } = await init()
-      myApp = app
-
-      let { server } = await startHttp( myApp, 0 )
-      app.on( 'appStarted', () => {
-        serverAddress = `http://localhost:${server.address().port}`
-      } )
-      testServer = server
+      let { app } = await beforeEachContext( );
+      ( { server, sendRequest } = await initializeTestServer( app ) )
 
       actor.id = await createUser( actor )
 
       testToken = `Bearer ${( await createPersonalAccessToken( actor.id, 'test token', [ 'users:invite' ] ) )}`
     } )
 
-    after( async() => {
-      await knex.migrate.rollback( )
-      if ( testServer )
-        testServer.close()
+    after( async ( ) => {
+      await server.close()
     } )
 
     it( 'should create a server invite', async() => {
@@ -235,10 +209,6 @@ describe( 'Server Invites @server-invites', ( ) => {
     } )
   } )
 } )
-
-function sendRequest( auth, obj, address = serverAddress ) {
-  return chai.request( address ).post( '/graphql' ).set( 'Authorization', auth ).send( obj )
-}
 
 const longInviteMessage =
   'Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur? At vero eos et accusamus et iusto odio dignissimos ducimus qui blanditiis praesentium voluptatum deleniti atque corrupti quos dolores et quas molestias excepturi sint occaecati cupiditate non provident, similique sunt in culpa qui officia deserunt mollitia animi, id est laborum et dolorum fuga. Et harum quidem rerum facilis est et expedita distinctio. Nam libero tempore, cum soluta nobis est eligendi optio cumque nihil impedit quo minus id quod maxime placeat facere possimus, omnis voluptas assumenda est, omnis dolor repellendus. Temporibus autem quibusdam et aut officiis debitis aut rerum necessitatibus saepe eveniet ut et voluptates repudiandae sint et molestiae non recusandae. Itaque earum rerum hic tenetur a sapiente delectus, ut aut reiciendis voluptatibus maiores alias consequatur aut perferendis doloribus asperiores repellat.'
